# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ByPass API",
        "description": "API de gestion de bypass réalisée par JOBS Conseil. Cette API permet de gérer les demandes de bypass, les équipements, les zones, les capteurs et les utilisateurs.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API ByPass"
        }
    ],
    "paths": {
...
        "/equipment/{equipment}/sensors": {
            "get": {
                "tags": [
                    "Capteurs"
                ],
                "summary": "Liste des capteurs d'un équipement",
                "description": "Récupère tous les capteurs associés à un équipement spécifique",
                "operationId": "096d320a08790a5c1f09db885d914343",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "ID de l'équipement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des capteurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sensor"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Capteurs"
                ],
                "summary": "Créer un capteur",
                "description": "Crée un nouveau capteur pour un équipement. Accessible uniquement aux administrateurs.",
                "operationId": "158d056440b3f7944bf51f4c2c1a96da",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "ID de l'équipement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "unit",
                                    "criticalThreshold"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Capteur Température"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Température"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "°C"
                                    },
                                    "criticalThreshold": {
                                        "type": "string",
                                        "example": "50"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "bypassed",
                                            "maintenance",
                                            "faulty",
                                            "calibration"
                                        ],
                                        "example": "active"
                                    },
                                    "last_reading": {
                                        "type": "number",
                                        "example": 25.5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Capteur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sensor"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "422": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sensors/{sensor}": {
            "get": {
                "tags": [
                    "Capteurs"
                ],
                "summary": "Détails d'un capteur",
                "description": "Récupère les détails d'un capteur spécifique avec son équipement",
                "operationId": "60dc5af5fb5d87ae03c74f6ffed53968",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "description": "ID du capteur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du capteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sensor"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Capteurs"
                ],
                "summary": "Mettre à jour un capteur",
                "description": "Met à jour les informations d'un capteur. Accessible uniquement aux administrateurs.",
                "operationId": "1d70ea6b51af8d33e8124a8e8eb06b4d",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "description": "ID du capteur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Capteur Température"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Température"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "°C"
                                    },
                                    "criticalThreshold": {
                                        "type": "string",
                                        "example": "50"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "bypassed",
                                            "maintenance",
                                            "faulty",
                                            "calibration"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Capteur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sensor"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Capteurs"
                ],
                "summary": "Supprimer un capteur",
                "description": "Supprime un capteur. Accessible uniquement aux administrateurs.",
                "operationId": "2cac48377a74621fd696b50a2dcb3218",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "description": "ID du capteur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Capteur supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Capteur supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sensors": {
            "get": {
                "tags": [
                    "Capteurs"
                ],
                "summary": "Liste de tous les capteurs",
                "description": "Récupère la liste de tous les capteurs avec leurs équipements et zones",
                "operationId": "7753a7bae22c1cf397f71a7c0f4354ab",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des capteurs paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Sensor"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "description": "Message d'erreur",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Sensor": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "equipment_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Capteur Température"
                    },
                    "type": {
                        "type": "string",
                        "example": "Température"
                    },
                    "code": {
                        "type": "string",
                        "example": "TEMP-001"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Token d'authentification Bearer (obtenu via /api/auth/login)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}